backend/requirements.txt
python-dotenv
psycopg2
pandas
Flask
flask-cors
scikit-learn
